# dockerize_lara_app/news_project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/news_project/public/api/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get User Preferences",
                "description": "Retrieve the user's preferences, including news sources, authors, and categories.",
                "operationId": "getPreferences",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Set User Preferences",
                "description": "Allows authenticated users to set their preferences for news sources, authors, and categories.",
                "operationId": "setPreferences",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "preferences"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The user's full name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "preferences": {
                                        "description": "An array of user preferences.",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "source",
                                                "author"
                                            ],
                                            "properties": {
                                                "source": {
                                                    "description": "The preferred news source",
                                                    "type": "string",
                                                    "example": "BBC News"
                                                },
                                                "author": {
                                                    "description": "The preferred author",
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "category_id": {
                                                    "description": "The preferred category ID (optional)",
                                                    "type": "integer",
                                                    "example": 1,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences set successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "preferences": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "source": {
                                                                    "type": "string",
                                                                    "example": "BBC News"
                                                                },
                                                                "author": {
                                                                    "type": "string",
                                                                    "example": "John Doe"
                                                                },
                                                                "category_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "property": null
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/news_project/public/api/personalized-feed": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get Personalized Feed",
                "description": "Retrieve a personalized feed of articles based on the user's preferences, including news sources, authors, and categories.",
                "operationId": "getPersonalizedFeed",
                "responses": {
                    "200": {
                        "description": "Personalized feed retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Preferences not set",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences not set"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}